+++ administrator/components/com_users/src/View/Notes/HtmlView.php
@@ -170,7 +170,7 @@ class HtmlView extends BaseHtmlView
 			}
 		}
 
-		if ($this->state->get('filter.published') == -2 && $canDo->get('core.delete'))
+		if (!$this->isEmptyState && $this->state->get('filter.published') == -2 && $canDo->get('core.delete'))
 		{
 			$toolbar->delete('notes.delete')
 				->text('JTOOLBAR_EMPTY_TRASH')
